+++ src/deluge/playback/mode/arrangement.cpp
@@ -351,6 +351,10 @@ void Arrangement::resetPlayPos(int32_t newPos, bool doingComplete, int32_t butto
 	playbackStartedAtPos = newPos;
 	lastProcessedPos = newPos;
 	arrangerView.xScrollWhenPlaybackStarted = currentSong->xScroll[NAVIGATION_ARRANGEMENT];
+	// if you were holding a clip pad and doing a reset,
+	// it can be easy to accidentally delete or enter the clip
+	// setting this to false prevents that
+	arrangerView.actionOnDepress = false;
 
 	if (currentSong->paramManager.mightContainAutomation()) {
 		char modelStackMemory[MODEL_STACK_MAX_SIZE];